% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM.R
\name{LM}
\alias{LM}
\title{LM}
\usage{
LM(formula, data, intercept = TRUE)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted. Note that the response variable should be numeric type.}

\item{data}{an optional data frame containing the variables in the model.}

\item{intercept}{If the model fit with intercept, intercept = TRUE;
if model fit without intercept, intercept = FALSE.
The default setting is intercept = TRUE.}
}
\value{
An object of class "LM" is a list containing at least the following components:

coefficients: a named vector of coefficients

residuals: the residuals, that is response minus fitted values.

rank: the numeric rank of the fitted linear model.

fitted.values: the fitted mean values.

df.residual: the residual degrees of freedom.

call: the matched call.

terms: the terms object used.

coefficients_df: the table includes estimates of beta coefficients, standard error, t value, and p-value

sigma: the residual standard error

r.squared: the proportion of the variance for a dependent variable that's explained by independent variable(s)

adj.r.squared: a penalized version of r.squared

fstatistic: the overall F statistic and its corresponding degrees of freedom of numerator and denominator

F.p-value: the overall F test p value

cov.unscaled: the unscaled var-cov matrix of estimates of beta coefficients
}
\description{
LM is used to fit linear models, including multivariate ones. It can be used to carry out regression.
It yields the same but combined output results from the build-in methods lm() and summary(lm()) (i.e. equivalent to some output which could be extracted with $ from lm() and summary(lm()) ).
}
\examples{
##Note: the examples are generated by R's default dataset iris

## fit the simple linear regression model
LM_result = LM(Sepal.Width ~ Sepal.Length, data = iris, intercept = FALSE) # omiting intercept
summaryLM(LM_result) ## obtain the printed summary

## fit the multiple linear regression model
LM_result = LM(Sepal.Width ~ Sepal.Length + Petal.Length, data = iris)
LM_result$coefficients ## obtain estimates of beta coefficients
LM_result$fitted.values ## obtain fitted mean values
LM_result$r.squared ## obtain r.squared

## fit the multiple linear regression model with interaction terms
LM_result = LM(Sepal.Width ~ Sepal.Length * Petal.Length, data = iris)
summaryLM(LM_result) ## obtain the printed summary

## fit the linear regression model with only interaction terms
LM_result = LM(Sepal.Width ~ I(Sepal.Length * Petal.Length), data = iris)
LM_result$coefficients_df ## obtain the table includes estimates of beta coefficients, standard error, t value, and p-value


}
