% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM.R
\name{LM}
\alias{LM}
\title{LM}
\usage{
LM(formula, data, intercept = TRUE)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}

\item{data}{an optional data frame containing the variables in the model.}

\item{intercept}{If the model fit with intercept, intercept = TRUE;
if model fit without intercept, intercept = FALSE.
The default setting is intercept = TRUE.}
}
\value{
a list of result:
coefficients, residuals, rank, fitted.values, df.residual, model, call, terms, coefficients_df, sigma, r.squared, adj.r.squared, fstatistic, F.p-value, cov.unscaled
}
\description{
LM is used to fit linear models, including multivariate ones. It can be used to carry out regression.
It yields the same results as the build-in methods lm() and summary(lm()).
}
\examples{
## fit the simple linear regression model
LM_result = LM(Sepal.Width ~ Sepal.Length, data = iris, intercept = FALSE) # omiting intercept

## fit the multiple linear regression model
LM_result = LM(Sepal.Width ~ Sepal.Length + Petal.Length, data = iris)
LM(Sepal.Width ~ Sepal.Length + Petal.Length, data = iris)$coefficients ## obtain coefficients

## fit the multiple linear regression model with interaction terms
LM(Sepal.Width ~ Sepal.Length * Petal.Length, data = iris)

## fit the linear regression model with only interaction terms
LM(Sepal.Width ~ I(Sepal.Length * Petal.Length), data = iris)



}
