% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM.R
\name{LM}
\alias{LM}
\title{LM}
\usage{
LM(formula, data)
}
\arguments{
\item{formula}{input the model which need to be fitted}

\item{data}{input the dataset that the fitted model used}
}
\value{
the a list of result: call, terms, residuals, coefficients, coefficients_table, fitted.values, sigma, rank,
df.residual, r.squared, adj.r.squared, fstatistic, model
}
\description{
Fit linear regression model by given dataset and corresponding interests of covariates
}
\examples{
LM(Sepal.Width ~ Sepal.Length, data = iris)
LM(Sepal.Width ~ Sepal.Length + Petal.Length, data = iris)
LM(Sepal.Width ~ Sepal.Length * Petal.Length, data = iris)
LM(Sepal.Width ~ I(Sepal.Length * Petal.Length), data = iris)

}
