% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryLM.R
\name{summaryLM}
\alias{summaryLM}
\title{summaryLM}
\usage{
summaryLM(result)
}
\arguments{
\item{result}{all output from fitted LM}
}
\description{
Mimic summary() of lm()
Print the summary of fitted model in LM()
}
