% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryLM.R
\name{summaryLM}
\alias{summaryLM}
\title{summaryLM}
\usage{
summaryLM(result)
}
\arguments{
\item{result}{the returned list of outputs by using LM() to fit the linear regression model}
}
\description{
summaryLM is a function used to print a summary of the results of the fitted model by using LM().
It mimics the printed output of the build-in method using summary(lm()).
}
\examples{
##Note: the examples are generated by default dataset iris

## fit the simple linear regression model
LM_result = LM(Sepal.Width ~ Sepal.Length, data = iris)
summaryLM(LM_result) ## obtain the summary of the fitted model

## fit the simple linear regression model
LM_result = LM(Sepal.Width ~ Sepal.Length, data = iris, intercept = FALSE) # omiting intercept
summaryLM(LM_result)

## fit the multiple linear regression model
LM_result = LM(Sepal.Width ~ Sepal.Length + Petal.Length, data = iris)
summaryLM(LM_result)

## fit the multiple linear regression model with interaction
LM_result = LM(Sepal.Width ~ Sepal.Length * Petal.Length, data = iris)
summaryLM(LM_result)

## fit the linear regression model with only interaction terms
LM_result = LM(Sepal.Width ~ I(Sepal.Length * Petal.Length), data = iris)
summaryLM(LM_result)

}
