% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryLM.R
\name{summaryLM}
\alias{summaryLM}
\title{summaryLM}
\usage{
summaryLM(result)
}
\arguments{
\item{result}{the summary of all output from fitted LM}
}
\description{
Print the summary of fitted model in LM
Mimic the printing output of the build-in method summary(lm())
}
\examples{
## fit the simple linear regression model
LM_result = LM(Sepal.Width ~ Sepal.Length, data = iris)
## obtain the summary of the fitted model
summaryLM(LM_result)

}
